+++ hdkeychain/doc.go
