+++ hdkeychain/extendedkey_test.go
@@ -17,7 +17,7 @@ import (
 	"testing"
 
 	"github.com/decred/dcrd/chaincfg"
-	"github.com/decred/dcrd/dcrutil/hdkeychain"
+	"github.com/decred/dcrd/hdkeychain"
 )
 
 // TestBIP0032Vectors tests the vectors provided by [BIP32] to ensure the
